Rename: Change filename extensions of components
