+++ package.json
@@ -3,6 +3,7 @@
   "version": "1.0.0-0",
   "private": true,
   "dependencies": {
+    "@coreui/icons-react": "^2.0.1",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
