--- src/components/LoadingSpinner.js
@@ -1,5 +1,0 @@
-const LoadingSpinner = () => {
-  return <div></div>;
-};
-
-export default LoadingSpinner;
