--- src/components/SwitchMode.js
@@ -1,11 +1,0 @@
-const SwitchMode = () => {
-  return (
-    <div className='SwitchMode'>
-      <div className='switch outer'>
-        <div className='switch inner'></div>
-      </div>
-    </div>
-  );
-};
-
-export default SwitchMode;
