+++ src/containers/date-view/MediaContainer.jsx
